MatrixPilot_fbw - Modified pyparam to intx_t types
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2362 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/nv_memory_table.c
@@ -18,7 +18,7 @@ const mavlink_parameter_block    mavlink_parameter_blocks[] = {
     };
 
 
-const unsigned int mavlink_parameter_block_count = sizeof(mavlink_parameter_blocks) / sizeof(mavlink_parameter_block);
+const uint16_t mavlink_parameter_block_count = sizeof(mavlink_parameter_blocks) / sizeof(mavlink_parameter_block);
 
 #endif    //USE_NV_MEMORY
